refactor: split BottomTabBar.tsx in hooks and utils
