+++ packages/bottom-tabs/src/utils/shouldUseHorizontalLabels.ts
@@ -1,0 +1,57 @@
+import type {
+  ParamListBase,
+  TabNavigationState,
+} from '@react-navigation/native';
+import { StyleSheet } from 'react-native';
+
+import type { BottomTabDescriptorMap } from '../types';
+
+const DEFAULT_MAX_TAB_ITEM_WIDTH = 125;
+
+export type ShouldUseHorizontalLabelsOptions = {
+  state: TabNavigationState<ParamListBase>;
+  descriptors: BottomTabDescriptorMap;
+  layout: { height: number; width: number };
+  dimensions: { height: number; width: number };
+};
+
+export default function shouldUseHorizontalLabels({
+  state,
+  descriptors,
+  layout,
+  dimensions,
+}: ShouldUseHorizontalLabelsOptions) {
+  const { tabBarLabelPosition } =
+    descriptors[state.routes[state.index].key].options;
+
+  if (tabBarLabelPosition) {
+    switch (tabBarLabelPosition) {
+      case 'beside-icon':
+        return true;
+      case 'below-icon':
+        return false;
+    }
+  }
+
+  if (layout.width >= 768) {
+    // Screen size matches a tablet
+    const maxTabWidth = state.routes.reduce((acc, route) => {
+      const { tabBarItemStyle } = descriptors[route.key].options;
+      const flattenedStyle = StyleSheet.flatten(tabBarItemStyle);
+
+      if (flattenedStyle) {
+        if (typeof flattenedStyle.width === 'number') {
+          return acc + flattenedStyle.width;
+        } else if (typeof flattenedStyle.maxWidth === 'number') {
+          return acc + flattenedStyle.maxWidth;
+        }
+      }
+
+      return acc + DEFAULT_MAX_TAB_ITEM_WIDTH;
+    }, 0);
+
+    return maxTabWidth <= layout.width;
+  } else {
+    return dimensions.width > dimensions.height;
+  }
+}
