+++ packages/bottom-tabs/src/views/BottomTabBar.tsx
@@ -3,126 +3,30 @@ import {
   CommonActions,
   NavigationContext,
   NavigationRouteContext,
-  ParamListBase,
-  TabNavigationState,
   useLinkBuilder,
   useTheme,
 } from '@react-navigation/native';
 import React from 'react';
 import {
   Animated,
-  LayoutChangeEvent,
   Platform,
   StyleProp,
   StyleSheet,
   View,
   ViewStyle,
 } from 'react-native';
-import { EdgeInsets, useSafeAreaFrame } from 'react-native-safe-area-context';
+import { useSafeAreaFrame } from 'react-native-safe-area-context';
 
-import type { BottomTabBarProps, BottomTabDescriptorMap } from '../types';
-import BottomTabBarHeightCallbackContext from '../utils/BottomTabBarHeightCallbackContext';
-import useIsKeyboardShown from '../utils/useIsKeyboardShown';
+import useTabBarAnimatedVisibility from '../hooks/useTabBarAnimatedVisibility';
+import useTabBarLayout from '../hooks/useTabBarLayout';
+import type { BottomTabBarProps, BottomTabNavigationOptions } from '../types';
+import getPaddingBottom from '../utils/getPaddingBottom';
+import getTabBarHeight from '../utils/getTabBarHeight';
+import shouldUseHorizontalLabels from '../utils/shouldUseHorizontalLabels';
 import BottomTabItem from './BottomTabItem';
 
 type Props = BottomTabBarProps & {
   style?: Animated.WithAnimatedValue<StyleProp<ViewStyle>>;
-};
-
-const DEFAULT_TABBAR_HEIGHT = 49;
-const COMPACT_TABBAR_HEIGHT = 32;
-const DEFAULT_MAX_TAB_ITEM_WIDTH = 125;
-
-const useNativeDriver = Platform.OS !== 'web';
-
-type Options = {
-  state: TabNavigationState<ParamListBase>;
-  descriptors: BottomTabDescriptorMap;
-  layout: { height: number; width: number };
-  dimensions: { height: number; width: number };
-};
-
-const shouldUseHorizontalLabels = ({
-  state,
-  descriptors,
-  layout,
-  dimensions,
-}: Options) => {
-  const { tabBarLabelPosition } =
-    descriptors[state.routes[state.index].key].options;
-
-  if (tabBarLabelPosition) {
-    switch (tabBarLabelPosition) {
-      case 'beside-icon':
-        return true;
-      case 'below-icon':
-        return false;
-    }
-  }
-
-  if (layout.width >= 768) {
-    // Screen size matches a tablet
-    const maxTabWidth = state.routes.reduce((acc, route) => {
-      const { tabBarItemStyle } = descriptors[route.key].options;
-      const flattenedStyle = StyleSheet.flatten(tabBarItemStyle);
-
-      if (flattenedStyle) {
-        if (typeof flattenedStyle.width === 'number') {
-          return acc + flattenedStyle.width;
-        } else if (typeof flattenedStyle.maxWidth === 'number') {
-          return acc + flattenedStyle.maxWidth;
-        }
-      }
-
-      return acc + DEFAULT_MAX_TAB_ITEM_WIDTH;
-    }, 0);
-
-    return maxTabWidth <= layout.width;
-  } else {
-    return dimensions.width > dimensions.height;
-  }
-};
-
-const getPaddingBottom = (insets: EdgeInsets) =>
-  Math.max(insets.bottom - Platform.select({ ios: 4, default: 0 }), 0);
-
-export const getTabBarHeight = ({
-  state,
-  descriptors,
-  dimensions,
-  insets,
-  style,
-  ...rest
-}: Options & {
-  insets: EdgeInsets;
-  style: Animated.WithAnimatedValue<StyleProp<ViewStyle>> | undefined;
-}) => {
-  // @ts-ignore
-  const customHeight = StyleSheet.flatten(style)?.height;
-
-  if (typeof customHeight === 'number') {
-    return customHeight;
-  }
-
-  const isLandscape = dimensions.width > dimensions.height;
-  const horizontalLabels = shouldUseHorizontalLabels({
-    state,
-    descriptors,
-    dimensions,
-    ...rest,
-  });
-  const paddingBottom = getPaddingBottom(insets);
-
-  if (
-    Platform.OS === 'ios' &&
-    !Platform.isPad &&
-    isLandscape &&
-    horizontalLabels
-  ) {
-    return COMPACT_TABBAR_HEIGHT + paddingBottom;
-  }
-
-  return DEFAULT_TABBAR_HEIGHT + paddingBottom;
 };
 
 export default function BottomTabBar({
@@ -135,7 +39,8 @@ export default function BottomTabBar({
   const { colors } = useTheme();
   const buildLink = useLinkBuilder();
 
-  const focusedRoute = state.routes[state.index];
+  const { routes, index } = state;
+  const focusedRoute = routes[index];
   const focusedDescriptor = descriptors[focusedRoute.key];
   const focusedOptions = focusedDescriptor.options;
 
@@ -152,87 +57,7 @@ export default function BottomTabBar({
   } = focusedOptions;
 
   const dimensions = useSafeAreaFrame();
-  const isKeyboardShown = useIsKeyboardShown();
-
-  const onHeightChange = React.useContext(BottomTabBarHeightCallbackContext);
-
-  const shouldShowTabBar = !(tabBarHideOnKeyboard && isKeyboardShown);
-
-  const visibilityAnimationConfigRef = React.useRef(
-    tabBarVisibilityAnimationConfig
-  );
-
-  React.useEffect(() => {
-    visibilityAnimationConfigRef.current = tabBarVisibilityAnimationConfig;
-  });
-
-  const [isTabBarHidden, setIsTabBarHidden] = React.useState(!shouldShowTabBar);
-
-  const [visible] = React.useState(
-    () => new Animated.Value(shouldShowTabBar ? 1 : 0)
-  );
-
-  React.useEffect(() => {
-    const visibilityAnimationConfig = visibilityAnimationConfigRef.current;
-
-    if (shouldShowTabBar) {
-      const animation =
-        visibilityAnimationConfig?.show?.animation === 'spring'
-          ? Animated.spring
-          : Animated.timing;
-
-      animation(visible, {
-        toValue: 1,
-        useNativeDriver,
-        duration: 250,
-        ...visibilityAnimationConfig?.show?.config,
-      }).start(({ finished }) => {
-        if (finished) {
-          setIsTabBarHidden(false);
-        }
-      });
-    } else {
-      setIsTabBarHidden(true);
-
-      const animation =
-        visibilityAnimationConfig?.hide?.animation === 'spring'
-          ? Animated.spring
-          : Animated.timing;
-
-      animation(visible, {
-        toValue: 0,
-        useNativeDriver,
-        duration: 200,
-        ...visibilityAnimationConfig?.hide?.config,
-      }).start();
-    }
-
-    return () => visible.stopAnimation();
-  }, [visible, shouldShowTabBar]);
-
-  const [layout, setLayout] = React.useState({
-    height: 0,
-    width: dimensions.width,
-  });
-
-  const handleLayout = (e: LayoutChangeEvent) => {
-    const { height, width } = e.nativeEvent.layout;
-
-    onHeightChange?.(height);
-
-    setLayout((layout) => {
-      if (height === layout.height && width === layout.width) {
-        return layout;
-      } else {
-        return {
-          height,
-          width,
-        };
-      }
-    });
-  };
-
-  const { routes } = state;
+  const [layout, handleLayout] = useTabBarLayout(dimensions);
 
   const paddingBottom = getPaddingBottom(insets);
   const tabBarHeight = getTabBarHeight({
@@ -242,6 +67,13 @@ export default function BottomTabBar({
     dimensions,
     layout,
     style: [tabBarStyle, style],
+  });
+
+  const { isTabBarHidden, animatedStyles } = useTabBarAnimatedVisibility({
+    tabBarHideOnKeyboard,
+    tabBarVisibilityAnimationConfig,
+    paddingBottom,
+    layout,
   });
 
   const hasHorizontalLabels = shouldUseHorizontalLabels({
@@ -262,22 +94,7 @@ export default function BottomTabBar({
             tabBarBackgroundElement != null ? 'transparent' : colors.card,
           borderTopColor: colors.border,
         },
-        {
-          transform: [
-            {
-              translateY: visible.interpolate({
-                inputRange: [0, 1],
-                outputRange: [
-                  layout.height + paddingBottom + StyleSheet.hairlineWidth,
-                  0,
-                ],
-              }),
-            },
-          ],
-          // Absolutely position the tab bar so that the content is below it
-          // This is needed to avoid gap at bottom when the tab bar is hidden
-          position: isTabBarHidden ? 'absolute' : (null as any),
-        },
+        animatedStyles,
         {
           height: tabBarHeight,
           paddingBottom,
@@ -318,19 +135,11 @@ export default function BottomTabBar({
             });
           };
 
-          const label =
-            options.tabBarLabel !== undefined
-              ? options.tabBarLabel
-              : options.title !== undefined
-              ? options.title
-              : route.name;
-
-          const accessibilityLabel =
-            options.tabBarAccessibilityLabel !== undefined
-              ? options.tabBarAccessibilityLabel
-              : typeof label === 'string' && Platform.OS === 'ios'
-              ? `${label}, tab, ${index + 1} of ${routes.length}`
-              : undefined;
+          const label = getLabel(options, route.name);
+          const accessibilityLabel = getAccessibilityLabel(
+            options,
+            `${label}, tab, ${index + 1} of ${routes.length}`
+          );
 
           return (
             <NavigationContext.Provider
@@ -389,3 +198,23 @@ const styles = StyleSheet.create({
     flexDirection: 'row',
   },
 });
+
+export const getLabel = (
+  { title, tabBarLabel }: BottomTabNavigationOptions,
+  defaultLabel: string
+) =>
+  tabBarLabel !== undefined
+    ? tabBarLabel
+    : title !== undefined
+    ? title
+    : defaultLabel;
+
+export const getAccessibilityLabel = (
+  { tabBarAccessibilityLabel }: BottomTabNavigationOptions,
+  defaultLabel: string
+) =>
+  tabBarAccessibilityLabel !== undefined
+    ? tabBarAccessibilityLabel
+    : Platform.OS === 'ios'
+    ? defaultLabel
+    : undefined;
